+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/Content.kt
@@ -24,8 +24,8 @@ import android.graphics.Bitmap
  * `Content` is composed of a one or more heterogeneous parts that can be represent data in
  * different formats, like text or images.
  *
- * @param role The producer of the content. Must be either 'user' or 'model'. By default, it's
- * "user".
+ * @param role The producer of the content. Must be either `"user"` or `"model"`. By default, it's
+ * `"user"`.
  * @param parts An ordered list of [Part] that constitute this content.
  */
 public class Content