+++ firebase-crashlytics-ndk/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+* [changed] Updated `firebase-crashlytics` dependency to v19.2.1
+
 
 # 19.2.0
 * [changed] Updated `firebase-crashlytics` dependency to v19.2.0