+++ firebase-inappmessaging/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 * [changed] Updated protobuf dependency to `3.25.5` to fix
-  [CVE-2024-7254](https://github.com/advisories/GHSA-735f-pc8j-v9w8).
+  [CVE-2024-7254](https://nvd.nist.gov/vuln/detail/CVE-2024-7254).
 
 # 21.0.0
 * [fixed] Fixed bad token exception while showing FIAM