+++ firebase-perf/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
+* [fixed] Fixed `IllegalStateException` that happened when starting a trace
+  before Firebase initializes.
 * [changed] Updated protobuf dependency to `3.25.5` to fix
-  [CVE-2024-7254](https://github.com/advisories/GHSA-735f-pc8j-v9w8).
+  [CVE-2024-7254](https://nvd.nist.gov/vuln/detail/CVE-2024-7254).
 
 # 21.0.1
 * [fixed] Fixed an `ExceptionInInitializerError` where the `url.openStream()` causes a crash if