+++ firebase-sessions/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unreleased
+* [changed] Updated protobuf dependency to `3.25.5` to fix
+  [CVE-2024-7254](https://github.com/advisories/GHSA-735f-pc8j-v9w8).
+
 
 # 2.0.5
 * [unchanged] Updated to keep SDK versions aligned.